Ensure JAR archive is reproducible
Having reproducible archives makes it easier for anyone who wants to
verify that the contents of the JAR is based on the published source
code since they can build the project themselves and compare the
checksums.

To achieve this we modify the archive so that the file order, file and
directory permissions, and timestamps are consistent across builds on
different machines. A different operating system, umask, and/or
version control system of choice might lead non-reproducibility if we
don't specify these explicitly.

In addition, we lock (transitive) dependency versions for the plugin
project and the build-script so that two subsequent builds don't run
the risk of getting different dependency versions resolved.