Activate Gradle build scans by default on CI
This commit activates build scans on all CI-builds by default which
are then subsequently published to the public Gradle build scan portal.

For convenience; An html-file, automatically redirecting the user to
the actual build scan, is created and uploaded as a build report for
each workflow run. This form of redirect will work in any (semi-)modern
browser.

To ensure local copies of previous build-scan.html files are removed on
`./gradlew clean` we apply the `base` plugin to the root project build
file.

In addition to information already gathered by default by gradle during
the scan, (named) tags are utilized to provide:

- A link back to the relevant branch or tag in the GitHub repository,
  for convenient navigation.
- Commit SHA, so that we can check out the appropriate revision locally
  when debugging/comparing locally.
- Which workflow and job that triggered the build in question, again
  for debugging purposes.

In accordance with GDPR, and basic courtesy really, personal
identifiable information, such as username, hostname and ip-addresses,
are removed when the environment variable 'CI' is not set. This will
allow people to create build scans without publishing their computer
names, IP's, etc.

I don't know what the value of the system property `user.name` would be
on the GitHub Actions runner, but to achieve a constant username across
CI-builds it's hardcoded to `github`.