Replace ./gradlew with gradle-build-action on CI
This simplifies running the build using different Gradle versions,
and it provides a more efficient caching of Gradle User Home between
workflow runs compared to setup-java.

GitHub's actions/cache is scoped to the key and branch, where the
default branch cache is available to other branches. This means
identical cache entries will be stored separately for different
branches. To reduce cache usage and cache eviction rate due to cache
size constraints, we only write to the cache on the 'main' branch.