Introduce intermediate env var for GitHub context
Store the value of `${{ github.event.pull_request.base.sha }}` in
memory and use it as an environment variable instead of interacting
with the script generation process and potentially opening the door
for script injection attacks.

I do not know off-hand if a GH Action or a run command can overwrite
the value of `${{ github.event.pull_request.base.sha }}`, but this will
at least follow best practices in dealing with GitHub contexts and
create a better example to follow even if it doesn't.

See also https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#understanding-the-risk-of-script-injections