Submit Gradle dependency graph to GitHub
The generated dependency graph is sent to GitHub on push to the default
branch as well as for pull requests. This allows Dependabot to suggest
updates and to give vulnerability alerts for direct and transitive
dependencies.

Due to security constraints on the CI server side the `contents: write`
permission is not available to a pull request from a publicly forked
repository. That being the case we need two separate workflows in order
to generate and submit the graph for all use-cases.

First we generate and upload the graph as a workflow artifact in the
main workflow, and then we trigger the second workflow to download the
graph and submit it. The second workflow (--submit-dependency-graph)
will only be triggered and run if the workflow file is on the default
branch according to the official documentation.