Cancel unnecessary workflow runs
There can be at most one running and one `pending` workflow run in the
same concurrency group at any time. When a new workflow run is queued
any existing `pending` workflow run is cancelled. Previously started
workflow runs in the same concurrency group is cancelled except for
when on a release branch.

When a workflow is running on a release branch we wish for it to finish
without interruption because there might be side effects such as
release artifacts created.

When a branch, tag or pull request receives a new push we should cancel
existing workflow runs and start new ones as the old run is outdated.
This will speed up the time it takes for the relevant workflow runs to
finish.

The value of `${{ github.workflow }}` is the value of the `name` field
in the workflow file. If the `name` field is not specified the value is
the full path of the file in the repository.

For workflows triggered by `push` the value of `${{ github.ref }}` is
the branch or tag ref that was pushed, such as `refs/heads/<branch_name>`
or `refs/tags/<tag_name>`. For workflows triggered by `pull_request`
the value is `refs/pull/<pr_number>/merge`. When triggered by
`workflow_dispatch` the value is the branch or tag that received
the dispatch such as `refs/heads/<branch_name>`. When triggered by
`schedule` the value is the default branch `refs/heads/main`.