Add deno dependency only if not already user defined
While not formally supported, configuring your own artifact repository
and dependency using the deno configuration would lead to having two
repositories defined as well as having the wrong installation directory
since the version part of the directory name was the one from the
extension and not the dependency. With this change the installation dir
is named correctly and the plugin will not add another repository, but
instead assumes that the buildscript using the plugin has defined one.